requirements.txt
langchain>=0.1.17
langchain-community>=0.0.34
langchain-core>=0.1.34
google-generativeai>=0.3.2
transformers
llama-cpp-python
faiss-cpu
psycopg2-binary
pymongo
streamlit
pytest
python-dotenv
PyYAML
tqdm
rich
